pyproject.toml
[project]
name = "flexiflow"
version = "0.1.0"
description = "A small async component engine with events, state machines, and a minimal CLI."
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "coverage>=7.0",
]
reload = ["watchfiles>=0.21"]
api = ["fastapi>=0.110", "uvicorn>=0.27"]

[project.scripts]
flexiflow = "flexiflow.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
